requirements.txt
streamlit
pandas
yfinance>=0.2.37
matplotlib
seaborn
feedparser
textblob
pandas-datareader
prophet
numpy
requests
curl-cffi
xlrd
selenium
webdriver-manager
html5lib
lxml
python-dotenv
pandas-ta
python-telegram-bot>=20.8
nest-asyncio
scipy
beautifulsoup4